+++ components/Explore/Category/Category.tsx
@@ -1,4 +1,4 @@
-import { Tag } from "@/components/ui/Tag/Tag";
+import { Tag } from "@/components/ui/Tag";
 
 interface ICategoryProps {
   category: string;
