+++ components/Forms/LoginForm/LoginForm.tsx
@@ -1,8 +1,8 @@
-import { Alert } from "@/components/ui/Alert/Alert";
-import { Button } from "@/components/ui/Button/Button";
-import { Label } from "@/components/ui/Label/Label";
-import { Input } from "@/components/ui/Input/Input";
-import { useLogin } from "@/hooks/useLogin/useLogin";
+import { Alert } from "@/components/ui/Alert";
+import { Button } from "@/components/ui/Button";
+import { Label } from "@/components/ui/Label";
+import { Input } from "@/components/ui/Input";
+import { useLogin } from "@/hooks/useLogin";
 
 interface ILoginFormProps {
   switchModal: () => void;
