+++ components/Forms/RegisterForm/RegisterForm.tsx
@@ -1,8 +1,8 @@
-import { Alert } from "@/components/ui/Alert/Alert";
-import { Button } from "@/components/ui/Button/Button";
-import { Label } from "@/components/ui/Label/Label";
-import { Input } from "@/components/ui/Input/Input";
-import { useRegister } from "@/hooks/useRegister/useRegister";
+import { Alert } from "@/components/ui/Alert";
+import { Button } from "@/components/ui/Button";
+import { Label } from "@/components/ui/Label";
+import { Input } from "@/components/ui/Input";
+import { useRegister } from "@/hooks/useRegister";
 
 interface IRegisterFormProps {
   switchModal: () => void;
