--- components/Home/Home/index.ts
@@ -1,1 +1,0 @@
-export { Home } from "./Home";
