+++ components/Home/index.ts
@@ -1,1 +1,1 @@
-export { Home } from "./Home/Home";
+export { Home } from "./Home";
