+++ components/Navbar/Navbar.tsx
@@ -1,8 +1,8 @@
 import { BiHeart, BiGlobe, BiLogOut, BiSlider } from "react-icons/bi";
-import { NavbarItem } from "./NavbarItem/NavbarItem";
-import { useModal } from "@/hooks/useModal/useModal";
+import { NavbarItem } from "./NavbarItem";
+import { useModal } from "@/hooks/useModal";
 import { LogoutModal } from "./LogoutModal";
-import { Logo } from "@/components/ui/Logo/Logo";
+import { Logo } from "@/components/ui/Logo";
 
 export const Navbar: React.FC = () => {
   const { isModalOpen, openModal, closeModal } = useModal();
