+++ components/Settings/Areas/Areas.tsx
@@ -1,10 +1,10 @@
 import type { Area } from "@/interfaces";
 import { isItemChosen } from "@/utils/isItemChosen";
-import { useSettingsForm } from "@/hooks/useSettingsForm/useSettingsForm";
-import { Checkbox } from "@/components/ui/Checkbox/Checkbox";
-import { Alert } from "@/components/ui/Alert/Alert";
-import { Button } from "@/components/ui/Button/Button";
-import { Label } from "@/components/ui/Label/Label";
+import { useSettingsForm } from "@/hooks/useSettingsForm";
+import { Checkbox } from "@/components/ui/Checkbox";
+import { Alert } from "@/components/ui/Alert";
+import { Button } from "@/components/ui/Button";
+import { Label } from "@/components/ui/Label";
 
 interface IAreasProps {
   allAreas: Area[];
