+++ components/Settings/Categories/Categories.tsx
@@ -1,10 +1,10 @@
 import type { Category } from "@/interfaces";
-import { useSettingsForm } from "@/hooks/useSettingsForm/useSettingsForm";
+import { useSettingsForm } from "@/hooks/useSettingsForm";
 import { isItemChosen } from "@/utils/isItemChosen";
-import { Checkbox } from "@/components/ui/Checkbox/Checkbox";
-import { Alert } from "@/components/ui/Alert/Alert";
-import { Button } from "@/components/ui/Button/Button";
-import { Label } from "@/components/ui/Label/Label";
+import { Checkbox } from "@/components/ui/Checkbox";
+import { Alert } from "@/components/ui/Alert";
+import { Button } from "@/components/ui/Button";
+import { Label } from "@/components/ui/Label";
 
 interface ICategoriesProps {
   allCategories: Category[];
