+++ components/UI/Notification/Notification.tsx
@@ -1,5 +1,5 @@
 import { motion } from "framer-motion";
-import { Portal } from "../Portal/Portal";
+import { Portal } from "../Portal";
 
 interface INotificationProps {
   children: React.ReactNode;
