+++ components/UI/Tag/index.ts
@@ -1,0 +1,1 @@
+export { Tag } from "./Tag";
