+++ styles/globals.css
@@ -39,25 +39,21 @@ a {
 @media (prefers-color-scheme: dark) {
 }
 
-/* Scroll bar stylings */
 ::-webkit-scrollbar {
   width: 5px;
   height: 5px;
 }
 
-/* Track */
 ::-webkit-scrollbar-track {
   background: #e7e7e7;
   border-radius: 5px;
 }
 
-/* Handle */
 ::-webkit-scrollbar-thumb {
   background: #888;
   border-radius: 5px;
 }
 
-/* Handle on hover */
 ::-webkit-scrollbar-thumb:hover {
   background: #787878;
 }
